+++ crates/canic-core/src/bootstrap/mod.rs
@@ -11,9 +11,10 @@ pub mod compiled {
         config::schema::{
             AppConfig, AppInitMode, AuthConfig, CanisterAuthConfig, CanisterConfig, CanisterKind,
             CanisterPool, ConfigModel, DelegatedTokenConfig, DirectoryConfig, DirectoryPool,
-            LogConfig, PoolImport, RandomnessConfig, RandomnessSource, RoleAttestationConfig,
-            ScalePool, ScalePoolPolicy, ScalingConfig, ShardPool, ShardPoolPolicy, ShardingConfig,
-            Standards, StandardsCanisterConfig, SubnetConfig, TopupPolicy, Whitelist,
+            FleetConfig, LogConfig, PoolImport, RandomnessConfig, RandomnessSource,
+            RoleAttestationConfig, ScalePool, ScalePoolPolicy, ScalingConfig, ShardPool,
+            ShardPoolPolicy, ShardingConfig, Standards, StandardsCanisterConfig, SubnetConfig,
+            TopupPolicy, Whitelist,
         },
         ids::{CanisterRole, SubnetRole},
     };